Refactor `InstanceInitializer` bean constellation.
* Combine `InstanceInitializer` and `InstanceInitializerBean` into a
single class, named `InstanceInitializer`.
* Remove `InstanceInitializerConfig` and instead make
`InstanceInitializer` the `@Component`.  We can undo this change if
we have a need to do extra configuration during initialization, but for
now it's not necessary.
* Adjust `InstanceInitializerTest` to clean up instances from DB both
before and after tests, as a test started failing due to what looked
like cross-test pollution after the aforementioned refactoring.
